Rough ability to add user-defined WFS servers.
