+++ server.js
@@ -22,7 +22,8 @@ var proxyAllowedHosts = {
     'spatialreference.org' : true,
     'www2.landgate.wa.gov.au' : true,
     'geofabric.bom.gov.au' : true,
-    'www.ga.gov.au' : true
+    'www.ga.gov.au' : true,
+    'data.gov.au' : true
 };
 
 app.get('/proxy', function(req, res) {
