+++ src/viewer/AusGlobeViewer.js
@@ -189,161 +189,9 @@ var AusGlobeViewer = function(geoDataManager) {
     this.viewer = undefined;
     this.map = undefined;
 
-    var categoryMapping = {
-        Layer : {
-            create : function(options) {
-                var parent = komapping.toJS(options.parent);
-                var data = combine(options.data, parent);
-
-                var layerViewModel = komapping.fromJS(data, categoryMapping);
-                layerViewModel.isEnabled = knockout.observable(false);
-
-                return layerViewModel;
-            }
-        }
-    };
-
-    var browserContentMapping = {
-        Layer : {
-            create : function(options) {
-                var layerViewModel = komapping.fromJS(options.data, categoryMapping);
-                layerViewModel.isOpen = knockout.observable(false);
-                layerViewModel.isLoading = knockout.observable(false);
-
-                if (!defined(layerViewModel.Layer)) {
-                    var layer;
-                    var layerRequested = false;
-                    var version = knockout.observable(0);
-
-                    layerViewModel.Layer = knockout.computed(function() {
-                        version();
-
-                        if (layerRequested) {
-                            return layer;
-                        }
-
-                        if (!defined(layer)) {
-                            layer = [];
-                        }
-
-                        // Don't request capabilities until the layer is opened.
-                        if (layerViewModel.isOpen()) {
-                            layerViewModel.isLoading(true);
-                            that.geoDataManager.getCapabilities(options.data, function(description) {
-                                var remapped = komapping.fromJS(description, categoryMapping);
-
-                                var layers = remapped.Layer();
-                                for (var i = 0; i < layers.length; ++i) {
-                                    // TODO: handle hierarchy better
-                                    if (defined(layers[i].Layer)) {
-                                        var subLayers = layers[i].Layer();
-                                        for (var j = 0; j < subLayers.length; ++j) {
-                                            layer.push(subLayers[j]);
-                                        }
-                                    } else {
-                                        layer.push(layers[i]);
-                                    }
-                                }
-
-                                version(version() + 1);
-                                layerViewModel.isLoading(false);
-                            });
-
-                            layerRequested = true;
-                        }
-
-                        return layer;
-                    });
-                }
-
-                return layerViewModel;
-            }
-        }
-    };
-
-    var browserContentViewModel = komapping.fromJS([], browserContentMapping);
-
-    var dataCollectionsPromise = loadJson('./data_collection.json');
-    var otherSourcesPromise = loadJson('./data_sources.json');
-
-    when.all([dataCollectionsPromise, otherSourcesPromise], function(sources) {
-        var browserContent = [];
-        browserContent.push(sources[0]);
-
-        var otherSources = sources[1].Layer;
-        for (var i = 0; i < otherSources.length; ++i) {
-            browserContent.push(otherSources[i]);
-        }
-
-        komapping.fromJS(browserContent, browserContentMapping, browserContentViewModel);
-    });
-
-    function noopHandler(evt) {
-        evt.stopPropagation();
-        evt.preventDefault();
-    }
-
-    function dropHandler(evt) {
-        evt.stopPropagation();
-        evt.preventDefault();
-
-        function loadCollection(json) {
-            if (!defined(json) || !defined(json.name) || !defined(json.Layer)) {
-                return;
-            }
-
-            var collections;
-            if (json.name === 'National Map Services') {
-                collections = json.Layer;
-            } else {
-                collections = [json];
-            }
-
-            var existingCollection;
-
-            for (var i = 0; i < collections.length; ++i) {
-                var collection = collections[i];
-
-                // Find an existing collection with the same name, if any.
-                var name = collection.name;
-                var existingCollections = browserContentViewModel();
-
-                existingCollection = undefined;
-                for (var j = 0; j < existingCollections.length; ++j) {
-                    if (existingCollections[j].name() === name) {
-                        existingCollection = existingCollections[j];
-                        break;
-                    }
-                }
-
-                if (defined(existingCollection)) {
-                    komapping.fromJS(collection, browserContentMapping, existingCollection);
-                } else {
-                    browserContentViewModel.push(komapping.fromJS(collection, browserContentMapping));
-                }
-            }
-        }
-
-        var files = evt.dataTransfer.files;
-        for (var i = 0; i < files.length; ++i) {
-            var file = files[i];
-            if (file.name.indexOf('.json') === -1) {
-                continue;
-            }
-
-            when(readJson(file), loadCollection);
-        }
-    }
-
-    document.addEventListener("dragenter", noopHandler, false);
-    document.addEventListener("dragexit", noopHandler, false);
-    document.addEventListener("dragover", noopHandler, false);
-    document.addEventListener("drop", dropHandler, false);
-
     this.geoDataBrowser = new GeoDataBrowser({
         viewer : this,
         container : leftArea,
-        content : browserContentViewModel,
         dataManager : geoDataManager
     });
